requirements.txt
anyio==3.6.1
base58==2.1.1
bson==0.5.10
certifi==2022.6.15
ECPy==1.2.5
fluent-logger==0.10.0
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
msgpack==1.0.4
prometheus-client==0.14.1
python-dateutil==2.8.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
typing-extensions==3.10.0.2
websockets==10.3
xrpl-py==1.6.0
aiohttp==3.8.3
aiofiles==23.1.0
